refactor: 拆分 test_rules.py 为每个 rules/*.py 对应一个测试文件
test_rules.py (1288行) → 8 个文件:
- test_node.py (298行): FormatNode 基类测试
- test_abstract.py (443行): 摘要相关节点测试
- test_heading.py (265行): 标题节点测试
- test_keywords.py (178行): 关键词节点测试
- test_body.py (194行): 正文节点测试
- test_caption.py (152行): 题注节点测试
- test_references.py (178行): 参考文献节点测试
- test_acknowledgement.py (202行): 致谢节点测试

934 passed, 0 failed
